Revert "rename utils.py"
This reverts commit 8918e17fabbab3fb1ded3e10624fd53c46407115.
