+++ process_video.py
@@ -2,7 +2,7 @@ import sys
 import requests
 
 from db import fetch_videos
-from nouns_utils import fetch_env_config
+from utils import fetch_env_config
 
 config = fetch_env_config()
